Fix enterFrame clearing timers too soon
Since the enterFrame function is the first to be run, it kept clearing the timer which had just fired, which in turn made it impossible to control this specific timer within its listener.

Now the enterFrame function will fire a timer's listener before removing the timer from runlist, which allows the user to control the timer from its listener.

-

table.insert and table.remove functions were also localised to improve their performance slightly.